--- src/components/editor/elements/text-placeholder.tsx
@@ -1,12 +1,0 @@
-import { TypeIcon } from "lucide-react";
-import Placeholder from "~/components/editor/elements/placeholder";
-
-type Props = {};
-
-export default function TextPlaceholder(props: Props) {
-  return (
-    <Placeholder type="text">
-      <TypeIcon size={40} className="text-muted-foreground" />
-    </Placeholder>
-  );
-}
