+++ src/components/editor/sidebar-elements-tab.tsx
@@ -1,8 +1,10 @@
 import React from "react";
-import ContainerPlaceholder from "~/components/editor/elements/container-placeholder";
-import MapPlaceholder from "~/components/editor/elements/map-placeholder";
-import TextPlaceholder from "~/components/editor/elements/text-placeholder";
-import TwoColumnsPlaceholder from "~/components/editor/elements/two-columns-placeholder";
+import {
+  ContainerPlaceholder,
+  MapPlaceholder,
+  TextPlaceholder,
+  TwoColumnsPlaceholder,
+} from "~/components/editor/placeholders";
 import type { EditorElementType } from "~/components/editor/type";
 import {
   Accordion,
